requirements.txt
streamlit
langchain_groq
langchain_community
langchain
langchain_core
deepgram-sdk
pydub
python-dotenv